+++ LoveBird/Sources/Network/APIClient.swift
@@ -22,7 +22,7 @@ public enum APIClient {
 
   // profile
   case fetchProfile
-  case editProfileAnnivarsary(image: UIImage?, editProfile: EditProfileAnnivarsaryRequest)
+  case editProfileAnnivarsary(image: Data?, profile: EditProfileAnnivarsaryRequest)
   case editProfile(image: Data?, profile: EditProfileRequest)
 
   // coupleLink
