+++ LoveBird/Sources/Network/LovebirdAPI.swift
@@ -26,6 +26,7 @@ protocol LovebirdAPIProtocol {
   // profile
   func fetchProfile() async throws -> Profile
   func editProfile(image: Data?, profile: EditProfileRequest) async throws -> Profile
+  func editProfileAnnivarsary(image: Data?, profile: EditProfileAnnivarsaryRequest) async throws -> Profile
 
   // coupleLink
   func linkCouple(linkCouple: LinkCoupleRequest) async throws -> StatusCode
@@ -125,5 +126,9 @@ struct LovebirdAPI: LovebirdAPIProtocol {
   
   func deleteSchedule(id: Int) async throws -> StatusCode {
     try await apiClient.requestRaw(.deleteSchedule(id: id)).status
+  }
+
+  func editProfileAnnivarsary(image: Data?, profile: EditProfileAnnivarsaryRequest) async throws -> Profile {
+    try await apiClient.request(.editProfileAnnivarsary(image: image, profile: profile))
   }
 }
