+++ LoveBird/Sources/Root/MyPage/MyPageEdit/MyPageProfileEditCore.swift
@@ -42,7 +42,7 @@ struct MyPageProfileEditCore: Reducer {
   
   enum Action: Equatable {
     case viewAppear
-    case backButtonTapped
+    case backTapped
     case isFocused(FocusedType)
     
     case nicknameEdited(String)
@@ -52,8 +52,8 @@ struct MyPageProfileEditCore: Reducer {
     
     case profileEditTapped
     case annivarsaryEditTapped
-    
-    case withdrawal
+    case withdrawalTapped
+
     case editProfileResponse(TaskResult<Profile>)
     case withdrawalResponse(TaskResult<String>)
     
@@ -65,119 +65,126 @@ struct MyPageProfileEditCore: Reducer {
     case anniversaryInitialized
     case birthdateInitialized
     case changeCurrentState(CurrentState)
+
+    case delegate(Delegate)
+    enum Delegate: Equatable {
+      case withdrawal
+    }
   }
   
-  @Dependency(\.apiClient) var apiClient
+  @Dependency(\.lovebirdApi) var lovebirdApi
   @Dependency(\.userData) var userData
-  
+
+  @Dependency(\.dismiss) var dismiss
+
   var body: some Reducer<State, Action> {
     Reduce { state, action in
       switch action {
-      case .viewAppear:
-        let profile = self.userData.get(key: .user, type: Profile.self)
-        if let profile { state.profile = profile }
-        return .none
-        
+
+      case .backTapped:
+        return .run { _ in await dismiss() }
+
       case .isFocused(let type):
         state.isNicknameFocused = type == .nickname
         state.isEmailFocused = type == .email
         state.isAnnivarsaryFocused = type == .annivarsary
         state.isBirthdateFocused = type == .birthdate
         return .none
-        
+
       case .nicknameEdited(let nickname):
         state.nickname = nickname
         return .none
-        
+
       case .emailEdited(let email):
         state.email = email
         return .none
-        
+
       case .annivarsaryEdited(let annivarsary):
         state.annivarsary = annivarsary
         return .none
-        
+
       case .birthdateEdited(let birthdate):
         state.birthdate = birthdate
         return .none
-        
+
       case .annivarsaryEditTapped:
         let request: EditProfileAnnivarsaryRequest = .init(
           annivarsary: state.annivarsary.toYMDFormat(),
           birthdate: state.birthdate.toYMDFormat()
         )
-        
+
         return .run { send in
           await send(
             .editProfileResponse(
               await TaskResult {
-                try await (self.apiClient.request(.editProfileAnnivarsary(
-                  image: nil,
-                  editProfile: request)
-                )) as Profile
+                try await lovebirdApi.editProfileAnnivarsary(image: nil, profile: request)
               }
             )
           )
         }
-        
+
       case .profileEditTapped:
         let request: EditProfileRequest = .init(
           nickname: state.nickname,
           email: state.email
         )
-        
+
         return .run { send in
           await send(
             .editProfileResponse(
               await TaskResult {
-                try await (self.apiClient.request(.editProfile(
-                  image: nil,
-                  editProfile: request)
-                )) as Profile
+                try await lovebirdApi.editProfile(image: nil, profile: request)
               }
             )
           )
         }
-        
-      case .withdrawal:
+
+      case .withdrawalTapped:
         return .run { send in
           await send(
             .withdrawalResponse(
               await TaskResult {
-                try await (self.apiClient.requestRaw(.withdrawal))
+                try await lovebirdApi.withdrawal()
               }
             )
           )
         }
-        
+
+      case .editProfileResponse(.success(let profile)):
+        self.userData.store(key: .user, value: profile)
+        return .run { _ in await dismiss() }
+
+      case .withdrawalResponse(.success):
+        return .send(.delegate(.withdrawal))
+
       case .showBottomSheet:
         state.showBottomSheet = true
         return .none
-        
+
       case .hideBottomSheet:
         state.showBottomSheet = false
         return .none
-        
+
       case .annivarsaryUpdated(let date):
         state.annivarsary = date
         return .none
-        
+
       case .birthdateUpdated(let date):
         state.birthdate = date
         return .none
-        
+
       case .anniversaryInitialized:
         state.annivarsary = .init()
         return .none
-        
+
       case .birthdateInitialized:
         state.birthdate = .init()
         return .none
-        
+
       case .changeCurrentState(let currentState):
         state.currentState = currentState
         return .none
-        
+
       default:
         return .none
       }
