+++ LoveBird/Sources/Root/MyPage/MyPageView.swift
@@ -88,58 +88,15 @@ struct MyPageView: View {
           .padding(.top, 30)
         
         VStack(spacing: 0) {
-          NavigationLinkStore(
-            self.store.scope(state: \.$myPageProfileEdit, action: MyPageAction.myPageProfileEdit)
-          ) {
-            viewStore.send(.editTapped)
-          } destination: { store in
-            MyPageProfileEditView(store: store)
-          } label: {
-            HStack(alignment: .center) {
-              Text("회원 정보 수정")
-                .font(.pretendard(size: 16))
-                .padding(.leading, 16)
-              
-              Spacer()
-            }
-            .frame(height: 68)
-            .frame(maxWidth: .infinity)
-            .bottomBorder()
+          Button { viewStore.send(.editTapped) } label: {
+            MyPageItemView(title: "프로필 수정")
           }
-          
+
           NavigationLink(destination: MyWebView(urlToLoad: Config.privacyPolicyURL)) {
-            HStack(alignment: .center) {
-              Text("FAQ")
-                .font(.pretendard(size: 16))
-                .padding(.leading, 16)
-              
-              Spacer()
-            }
-            .frame(height: 68)
-            .frame(maxWidth: .infinity)
-            .bottomBorder()
+            MyPageItemView(title: "개인정보 처리방침")
           }
-          
-          NavigationLink(destination: MyWebView(urlToLoad: Config.privacyPolicyURL)) {
-            HStack(alignment: .center) {
-              Text("개인정보 처리방침")
-                .font(.pretendard(size: 16))
-                .padding(.leading, 16)
-              
-              Spacer()
-            }
-            .frame(height: 68)
-            .frame(maxWidth: .infinity)
-            .bottomBorder()
-          }
-          
-          HStack(alignment: .center) {
-            Text("버전정보")
-              .font(.pretendard(size: 16))
-              .padding(.leading, 16)
-            
-            Spacer()
-            
+
+          MyPageItemView(title: "버전정보") {
             Text(Bundle.main.version)
               .font(.pretendard(size: 16))
               .padding(.trailing, 16)
